# specs/openapi.yaml
openapi: 3.1.0
info:
  title: App Gateway Authentication & Routing Service
  version: 1.0.0
  description: |
    The App Gateway is the unified authentication and request-routing gateway for the multi-app
    platform. It provides:

    - **JWT-based authentication** with short-lived access tokens (15 min) and rotating refresh
      tokens (7 days). Users authenticate once and stay authenticated across web and mobile clients.
    - **Enterprise SSO via SAML 2.0**, supporting Okta, Azure Active Directory, and Google
      Workspace as Identity Providers with per-tenant configuration.
    - **OAuth 2.0 Authorization Server** implementing `authorization_code` (with mandatory PKCE
      for public clients) and `client_credentials` grant types.
    - **Request proxying** to upstream backend services with authenticated identity forwarded
      via `X-User-*` headers. Upstream services require no token validation logic.
    - **Rate limiting**, **token revocation** (Redis-backed), **JWKS publication**, and full
      structured audit logging.

    ## Authentication

    Most endpoints that return or validate tokens use one of three security schemes:

    | Scheme       | Usage                                                         |
    |--------------|---------------------------------------------------------------|
    | `BearerAuth` | Authenticated user or service sends `Authorization: Bearer <jwt>` |
...

    ## Error Format

    All error responses use the `ErrorResponse` schema:
...

...
servers:
...
paths:
...
  /v1/auth/login:
    post:
      summary: Authenticate user with email and password
      description: |
        Validates the provided email and password against the `users` table using
        bcrypt comparison (minimum cost factor 12). On success, issues a signed RS256
        JWT access token (15-minute TTL) and a 256-bit URL-safe refresh token
        (7-day TTL). The plaintext refresh token is returned once; only its SHA-256
        hash is stored in the database.

        An `audit_log` entry is written for both successful and failed attempts.
        The endpoint is rate-limited to **100 requests per minute per source IP**.
      operationId: loginUser
      tags:
        - Authentication
      x-rate-limit:
        requests: 100
        window: 60
        per: ip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: alice@example.com
              password: "S3cur3P@ssw0rd!"
      responses:
        '200':
          description: Authentication successful. Returns JWT access token and refresh token.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              example:
                access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImtleS0yMDI2MDEifQ.eyJzdWIiOiI1NTBlODQwMC1lMjliLTQxZDQtYTcxNi00NDY2NTU0NDAwMDAiLCJlbWFpbCI6ImFsaWNlQGV4YW1wbGUuY29tIiwicm9sZXMiOlsidmlld2VyIl0sImF1dGhfbWV0aG9kIjoicGFzc3dvcmQiLCJ0ZW5hbnRfaWQiOm51bGwsInNjb3BlIjoib3BlbmlkIHByb2ZpbGUgZW1haWwgYXBpOnJlYWQiLCJpYXQiOjE3NTM2NTEwMDAsImV4cCI6MTc1MzY1MTkwMCwibmJmIjoxNzUzNjUxMDAwLCJpc3MiOiJodHRwczovL2dhdGV3YXkuZXhhbXBsZS5jb20iLCJhdWQiOiJwbGF0Zm9ybSIsImp0aSI6ImI0N2FjMTBiLTU4Y2MtNGM5Mi04MDA5LWRhZjU0Y2E0NzM3NCJ9.signature"
                refresh_token: "dGhpcyBpcyBhIDI1Ni1iaXQgdXJsLXNhZmUgcmFuZG9tIHN0cmluZw"
                token_type: Bearer
                expires_in: 900
                scope: "openid profile email api:read"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid_credentials
                error_description: Email or password is incorrect.
                request_id: "b47ac10b-58cc-4c92-8009-daf54ca47374"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/refresh:
    post:
      summary: Rotate refresh token and issue new token pair
      description: |
        Accepts a valid plaintext refresh token, locates the matching record by
        SHA-256 hash, verifies it is not expired or revoked, and issues a new
        access token + refresh token pair. The consumed refresh token is marked
        revoked in the database atomically.

        **Replay Attack Detection:** If the presented refresh token has already
        been used (record is already revoked), the gateway immediately revokes
        **all** refresh tokens in the same `session_family`, forcing full
        re-authentication for all sessions associated with that token lineage.

        **Idempotent Retries:** An optional client-generated `idempotency_key`
        (UUID) may be included. If a request presents an `idempotency_key`
        matching one seen within the last 30 seconds for the same
        `session_family`, the gateway returns the previously issued token pair
        unchanged instead of re-running rotation or reuse-detection logic —
        this lets a client safely retry after a lost response without
        triggering replay-attack detection. Requests without a matching cached
        `idempotency_key` follow normal rotation/reuse-detection behavior.

        Rate-limited to **100 requests per minute per source IP**.
      operationId: refreshToken
      tags:
        - Authentication
      x-rate-limit:
        requests: 100
        window: 60
        per: ip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refresh_token: "dGhpcyBpcyBhIDI1Ni1iaXQgdXJsLXNhZmUgcmFuZG9tIHN0cmluZw"
      responses:
        '200':
          description: Token rotation successful. Returns new access token and refresh token.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Refresh token is invalid, expired, or has already been used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expired:
                  summary: Expired refresh token
                  value:
                    error: token_expired
                    error_description: The refresh token has expired. Please log in again.
                    request_id: "b47ac10b-58cc-4c92-8009-daf54ca47374"
                reuse:
                  summary: Replay / reuse detected
                  value:
                    error: token_reuse_detected
                    error_description: Refresh token reuse detected. All sessions have been revoked.
                    request_id: "c58bd21c-69dd-5da3-9110-ebg65db58485"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/logout:
    post:
      summary: Revoke access token and refresh token (logout)
      description: |
        Revokes the authenticated session. The access token's `jti` is added to
        the Redis revocation set with a TTL equal to the token's remaining lifetime.
        The associated refresh token is marked revoked in PostgreSQL.

        An optional `refresh_token` body parameter may be provided to target a
        specific refresh token; if omitted, the most-recently-issued refresh token
        for the session is revoked.

        Returns `204 No Content` on success. Subsequent use of the revoked access
        token returns `401 Unauthorized`.
      operationId: logoutUser
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logout successful. Tokens are revoked.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/me:
    get:
      summary: Get authenticated user profile
      description: |
        Returns the authenticated user's profile, derived from the validated JWT
        access token claims. No additional database query is performed; all fields
        are sourced from the token payload. This endpoint is useful for clients
        to hydrate their UI immediately after login without a separate profile API call.
      operationId: getMe
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Returns the authenticated user's profile.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                email: alice@example.com
                first_name: Alice
                last_name: Smith
                roles:
                  - viewer
                  - api_user
                auth_method: password
                tenant_id: null
                email_verified: true
                created_at: "2026-01-15T09:00:00.000Z"
                last_login_at: "2026-07-27T14:30:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/.well-known/jwks.json:
    get:
      summary: Retrieve JSON Web Key Set (JWKS)
      description: |
        Returns the gateway's active RSA public keys in JWK format (RFC 7517).
        Upstream services and OAuth clients use this endpoint to verify JWT
        signatures without contacting the gateway for every request.

        The response includes all currently active and recently-rotated key IDs
        to support zero-downtime key rotation. New tokens are signed with the
        most-recently-activated key; old keys remain in the set until all tokens
        signed with them have naturally expired (max 15 minutes for access tokens).

        **Caching:** Responses are cached with `Cache-Control: public, max-age=3600`.
        The endpoint is served from in-process memory and responds in < 50 ms.
        No authentication is required.
      operationId: getJwks
      tags:
        - Authentication
      security: []
      responses:
        '200':
          description: JSON Web Key Set containing active public keys.
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=3600
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwksResponse'
              example:
                keys:
                  - kty: RSA
                    use: sig
                    kid: "key-20260101"
                    alg: RS256
                    "n": "sn38TJguMmTaTBm3_B4VdnXbP..."
                    "e": AQAB
                  - kty: RSA
                    use: sig
                    kid: "key-20260701"
                    alg: RS256
                    "n": "yq49UKhlNnTbCm4_C5WeoBz..."
                    "e": AQAB
        '500':
          $ref: '#/components/responses/InternalServerError'

  # -------------------------------------------------------------------------
  # SSO — SAML 2.0
  # -------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token issued by the gateway's `/v1/auth/login`,
        `/v1/auth/refresh`, or `/v1/oauth/token` endpoints. Tokens are
        RS256-signed and expire after 15 minutes (900 seconds). Include in the
        `Authorization` header as `Bearer <token>`.
    BasicAuth:
      type: http
      scheme: basic
      description: |
        OAuth 2.0 client authentication. The username is `client_id` and the
        password is `client_secret`. Used on the `/v1/oauth/token`,
        `/v1/oauth/introspect`, and `/v1/oauth/revoke` endpoints.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
      description: |
        Admin API key with `admin:oauth` scope. Required for admin-only
        operations such as OAuth client registration. Keys are managed
        out-of-band by the platform operations team.

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  headers:
    X-Request-Id:
      description: UUID v4 request identifier for distributed tracing.
      schema:
        type: string
        format: uuid
        example: "b47ac10b-58cc-4c92-8009-daf54ca47374"
    X-RateLimit-Limit:
      description: The maximum number of requests allowed in the current window.
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: The number of requests remaining in the current window.
      schema:
        type: integer
        example: 73
    X-RateLimit-Reset:
      description: Unix epoch timestamp (seconds) when the current rate limit window resets.
      schema:
        type: integer
        example: 1753651060
    Retry-After:
      description: |
        Number of seconds to wait before retrying (present on 429 responses only).
      schema:
        type: integer
        example: 47

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: The user's registered email address.
          example: alice@example.com
          maxLength: 254
        password:
          type: string
          format: password
          description: The user's password (plaintext; TLS required).
          example: "S3cur3P@ssw0rd!"
          minLength: 8
          maxLength: 128
      additionalProperties: false
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: |
            The plaintext refresh token received from a previous login or
            token refresh response.
          minLength: 32
          maxLength: 512
          example: "dGhpcyBpcyBhIDI1Ni1iaXQgdXJsLXNhZmUgcmFuZG9tIHN0cmluZw"
        idempotency_key:
          type: string
          format: uuid
          description: |
            Optional client-generated UUID. A retry presenting the same
            idempotency_key within a 30-second window returns the original
            token pair instead of re-running rotation/reuse-detection logic.
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      additionalProperties: false
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: |
            Optional. If provided, this specific refresh token is revoked.
            If omitted, the most-recently-issued refresh token for the
            session is revoked.
          example: "dGhpcyBpcyBhIDI1Ni1iaXQgdXJsLXNhZmUgcmFuZG9tIHN0cmluZw"
      additionalProperties: false

    # --- SAML Request Bodies ---
...
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - scope
      properties:
        access_token:
          type: string
          description: |
            Signed RS256 JWT access token. Expires in `expires_in` seconds.
            Decode the payload to read `sub`, `email`, `roles`, `auth_method`,
            `scope`, and other standard claims.
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImtleS0yMDI2MDEifQ..."
        refresh_token:
          type: string
          nullable: true
          description: |
            Rotating refresh token (256-bit URL-safe random string). Present
            for `authorization_code` and `password` grants when `offline_access`
            scope is granted. Absent for `client_credentials` grant.
            Expires in 7 days. Must be stored securely.
          example: "dGhpcyBpcyBhIDI1Ni1iaXQgdXJsLXNhZmUgcmFuZG9tIHN0cmluZw"
        token_type:
          type: string
          enum:
            - Bearer
          description: Always `"Bearer"`.
          example: Bearer
        expires_in:
          type: integer
          description: |
            Lifetime of the access token in seconds from issuance.
            `900` for user tokens (15 min); `3600` for client_credentials tokens.
          example: 900
        scope:
          type: string
          description: |
            Space-separated list of scopes granted. May be a subset of the
            requested scopes if the server downscoped the token.
          example: "openid profile email api:read offline_access"
      additionalProperties: false
...
    JwksResponse:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          description: Array of JSON Web Keys. May contain multiple keys for rotation support.
          items:
            $ref: '#/components/schemas/JwkKey'
          minItems: 1
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: |
            Machine-readable error code. Common values:
            - `invalid_credentials` — wrong email or password
            - `token_expired` — JWT or refresh token past its expiry
            - `token_revoked` — token has been explicitly revoked
            - `token_reuse_detected` — refresh token replay attack detected
            - `unauthorized` — missing or invalid Bearer token
            - `forbidden` — authenticated but not authorized
            - `insufficient_scope` — token lacks required OAuth scope
            - `tenant_not_found` — SAML tenant slug not found
            - `saml_validation_failed` — SAML assertion validation error
            - `route_not_found` — no proxy route configured for the path
            - `bad_gateway` — upstream service unreachable
            - `rate_limit_exceeded` — too many requests
            - `internal_server_error` — unexpected server-side error
          example: unauthorized
        error_description:
          type: string
          description: Human-readable explanation of the error. Not for programmatic use.
          example: "The access token is missing or invalid."
        error_uri:
          type: string
          format: uri
          description: Optional URI to documentation about this error type.
          example: "https://platform.example.com/docs/errors#unauthorized"
        request_id:
          type: string
          format: uuid
          description: |
            The `X-Request-Id` for this request. Include in support tickets to
            enable log correlation.
          example: "b47ac10b-58cc-4c92-8009-daf54ca47374"
      additionalProperties: false
...
  responses:
    BadRequest:
      description: |
        The request was malformed or failed input validation. The `error_description`
        field provides details on which parameter(s) were invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: invalid_request
            error_description: "Request body validation failed: 'email' must be a valid email address."
            request_id: "b47ac10b-58cc-4c92-8009-daf54ca47374"
    Unauthorized:
      description: |
        Authentication is required and has failed, or the provided credentials
        are missing or invalid. Check the `error` field for the specific reason.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: 'Bearer realm="app-gateway", error="invalid_token"'
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: unauthorized
            error_description: "A valid Bearer token is required."
            request_id: "b47ac10b-58cc-4c92-8009-daf54ca47374"
...
    TooManyRequests:
      description: |
        The rate limit for this client/IP has been exceeded. Wait until the
        `X-RateLimit-Reset` timestamp or the `Retry-After` seconds have elapsed
        before retrying.
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: rate_limit_exceeded
            error_description: "Rate limit of 100 requests per minute exceeded. Retry after 47 seconds."
            request_id: "b47ac10b-58cc-4c92-8009-daf54ca47374"
    InternalServerError:
      description: |
        An unexpected error occurred on the server. The `request_id` can be
        used to correlate this error with server-side logs.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: internal_server_error
            error_description: "An unexpected error occurred. Please try again or contact support."
            request_id: "b47ac10b-58cc-4c92-8009-daf54ca47374"